Refector the Atom monad to allow selective rewriting depending on compilation backend
The Atom monad is now a Reader/Writer/State monad. The reader contains a
compilation context that allows swtiching between different
interpretations of the atom combinators at compile time (see 'clocked'
for example). The writer maintains a list of compiler warnings. The
state contains the original Atom monad state.

The 'period' and 'phase' functions should now be deprecated in favor of
the 'clocked' combinator which takes into account the compilation
backend.